+++ lib/in_app_configs.dart
@@ -1,2 +1,2 @@
+export 'src/builder.dart';
 export 'src/configs.dart';
-export 'src/delegate.dart';