+++ lib/src/delegate.dart
@@ -1,3 +1,0 @@
-import 'package:in_app_remote/remote.dart';
-
-abstract class ConfigsDelegate extends RemoteDelegate {}